Replace Transformer with Sanitizer
Tried to keep them separate but it was starting to get too messy. Users can just
pass in custom transformers in the second constructor argument. Plus, this more
closely matches the ruby sanitize API.

Changed config to be a struct. So far, have only made a `DEFAULT_CONFIG`. Need
to make all of the others as well.

Started working on a `remove_contents_when_unwrapped` option to the config that
will indicate whether to remove contents of elements that are unwrapped.